# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TaskEither_any.any_": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"BodyTypeOf_typeofauthDto.input.register_": {
				"properties": {
					"lastname": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"birthday": {
						"type": "string",
						"format": "date-time"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"BodyTypeOf_typeofauthDto.input.login_": {
				"properties": {
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"BodyTypeOf_typeofauthDto.input.confirm_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "sample-fpts-v2",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskEither_any.any_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"username": "username",
											"firstname": "firstname",
											"lastname": "lastname",
											"email": "email@email.email",
											"createdAt": "2023-01-14T16:26:57.587Z"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BodyTypeOf_typeofauthDto.input.register_"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskEither_any.any_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"jwt": "true"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BodyTypeOf_typeofauthDto.input.login_"
							}
						}
					}
				}
			}
		},
		"/auth/confirmation": {
			"post": {
				"operationId": "Confirmation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskEither_any.any_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "63c25f4114226535d99056dd",
											"username": "username",
											"email": "email@email.email",
											"birthday": "2023-01-14T16:26:57.598Z",
											"country": "Iran",
											"firstname": "firstname",
											"lastname": "lastname"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BodyTypeOf_typeofauthDto.input.confirm_"
							}
						}
					}
				}
			}
		},
		"/user": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskEither_any.any_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"count": 1,
											"values": [
												{
													"id": "63bea0fa1bc80f6d62334fb4",
													"username": "username",
													"email": "email@email.email"
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "num",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/{userId}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskEither_any.any_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"user": {
												"id": "63bea0fa1bc80f6d62334fb4",
												"username": "username",
												"email": "email@email.email",
												"firstname": "firstname",
												"lastname": "lastname",
												"birthday": "2023-01-14T16:26:57.620Z",
												"country": "Iran"
											}
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}